Silence compiler warnings from regression tests.
Three categories of warnings:

1. Missing return at end of main() function
2. Unused variables (copy-pasted from another test)
3. Data type mismatch in printfs, e.g. passing SQLULEN param to %d

+++ test/src/arraybinding-test.c
@@ -175,8 +175,8 @@ int main(int argc, char **argv)
 		}
 	}
 
-	printf ("Fetching result sets for array bound (%d results expected)\n",
-			nprocessed);
+	printf ("Fetching result sets for array bound (%u results expected)\n",
+			(unsigned int) nprocessed);
 	for (i = 1; rc == SQL_SUCCESS || rc == SQL_SUCCESS_WITH_INFO; i++)
 	{
 		printf("%d: ", i);
